requirements.txt
numpy>=1.26
PyYAML>=6.0
tqdm>=4.66
torch>=2.3
onnx>=1.16
onnxruntime>=1.20
onnxruntime-tools>=1.7.0
